Внесены изменения в структуру приложения.
