+++ TestTask/Helpers/CharType.cs
@@ -1,0 +1,18 @@
+﻿namespace TestTask.Helpers
+{
+    /// <summary>
+    /// Тип букв.
+    /// </summary>
+    public enum CharType
+    {
+        /// <summary>
+        /// Гласные.
+        /// </summary>
+        Vowel,
+
+        /// <summary>
+        /// Согласные.
+        /// </summary>
+        Consonants
+    }
+}
