+++ TestTask/Helpers/CharTypeHelper.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TestTask.Helpers
+{
+    internal class CharTypeHelper
+    {
+        /// <summary>
+        /// Набор гласных букв для русского языка.
+        /// </summary>
+        internal const string RUS_VOWELS = "АУОЫИЭЯЮЁЕ";
+
+        /// <summary>
+        /// Набор согласных букв для русского языка.
+        /// </summary>
+        internal const string RUS_CONSONANTS = "БВГДЖЗЙКЛМНПРСТФХЦЧШЩ";
+
+        /// <summary>
+        /// Набор гласных букв для английского языка.
+        /// </summary>
+        internal const string EN_VOWELS = "AEIOUY";
+
+        /// <summary>
+        /// Набор согласных букв для английского языка.
+        /// </summary>
+        internal const string EN_CONSONANTS = "BCDFGHJKLMNPQRSTVWXZ";
+    }
+}
