+++ TestTask/Interfaces/IReadOnlyStream.cs
@@ -1,0 +1,32 @@
+﻿using System;
+
+namespace TestTask.Interfaces
+{
+    /// <summary>
+    /// Интерфейс для работы с файлом в сильно урезаном виде.
+    /// Умеет всего 2 вещи: прочитать символ, и перемотать стрим на начало.
+    /// </summary>
+    internal interface IReadOnlyStream : IDisposable
+    {
+        // TODO : Необходимо доработать данный интерфейс для обеспечения гарантированного закрытия файла, по окончанию работы с таковым!
+
+        /// <summary>
+        /// Метод чтения следующего символа из потока.
+        /// Если произведена попытка прочитать символ после достижения конца файла, метод 
+        /// должен бросать соответствующее исключение
+        /// </summary>
+        /// <returns>Считанный символ.</returns>
+        char ReadNextChar();
+
+        /// <summary>
+        /// Сбрасывает текущую позицию потока на начало.
+        /// </summary>
+        void ResetPositionToStart();
+
+        /// <summary>
+        /// Флаг окончания файла.
+        /// </summary>
+        bool IsEndOfFile { get; }
+
+    }
+}
