+++ TestTask/TestTask.csproj
@@ -42,12 +42,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CharType.cs" />
-    <Compile Include="IReadOnlyStream.cs" />
-    <Compile Include="LetterStats.cs" />
+    <Compile Include="Helpers\CharType.cs" />
+    <Compile Include="Helpers\CharTypeHelper.cs" />
+    <Compile Include="Interfaces\IReadOnlyStream.cs" />
+    <Compile Include="Models\LetterStats.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ReadOnlyStream.cs" />
+    <Compile Include="Models\ReadOnlyStream.cs" />
+    <Compile Include="Services\TextStatisticService.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
